separate point tests in other file
